requirements.txt
asgiref==3.3.1
Django==3.1.6
django-allauth==0.44.0
django-crispy-forms==1.11.1
django-tinymce==3.2.0
oauthlib==3.1.0
Pillow==8.1.1
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
